chore : Split FE, BE
