+++ package.json
@@ -3,19 +3,28 @@
   "private": true,
   "version": "0.0.0",
   "type": "module",
+  "description": "### [🐼판다마켓](https://been-panda.vercel.app) : https://been-panda.vercel.app",
+  "main": "vite.config.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/De-cal/6-sprint-mission-fe.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/De-cal/6-sprint-mission-fe/issues"
+  },
+  "homepage": "https://github.com/De-cal/6-sprint-mission-fe#readme",
   "scripts": {
-    "node": "nodemon src/backend/app.js",
-    "start": "node src/backend/app.js",
     "dev": "vite",
     "build": "vite build",
     "lint": "eslint .",
     "preview": "vite preview"
   },
   "dependencies": {
+    "@prisma/client": "^6.5.0",
     "axios": "^1.8.1",
-    "cors": "^2.8.5",
-    "express": "^4.21.2",
-    "mongoose": "^8.12.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-helmet": "^6.1.0",
@@ -24,16 +33,17 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.21.0",
+    "@types/node": "^22.13.10",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
     "@vitejs/plugin-react": "^4.3.4",
-    "dotenv": "^16.4.7",
     "eslint": "^9.21.0",
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^15.15.0",
-    "nodemon": "^3.1.9",
+    "tsx": "^4.19.3",
+    "typescript": "^5.8.2",
     "vite": "^6.2.0"
   }
 }
