--- src/backend/model/productSchema.js
@@ -1,37 +1,0 @@
-import mongoose from "mongoose";
-
-const ProductSchema = new mongoose.Schema(
-  {
-    name: {
-      type: String,
-      minLength: 1,
-      maxLength: 10,
-      required: true,
-    },
-    description: {
-      type: String,
-      minLength: 10,
-      maxLength: 100,
-      required: true,
-    },
-    price: {
-      type: Number,
-      minLength: 1,
-      required: true,
-    },
-    tags: {
-      type: [String],
-      maxLength: 5,
-      required: true,
-    },
-    like: {
-      type: Number,
-      default: 10,
-    },
-  },
-  {
-    timestamps: true,
-  }
-);
-
-export const Product = mongoose.model("Product", ProductSchema);
